backend/requirements-prod.txt
# ============================================================================
# PRODUCTION - Versiones PINNEADAS para build rápido y reproducible
# ============================================================================

# Core LangChain (versiones exactas para evitar resolution-too-deep)
langchain==0.3.7
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-community==0.3.5
langchain-chroma==0.1.4

# FastAPI Stack
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-core==2.23.4
python-multipart==0.0.17
sse-starlette==2.2.1
httpx==0.27.2

# Ollama
ollama==0.3.3

# Vector Databases
chromadb==0.5.18
faiss-cpu==1.9.0

# Document Processing
pypdf==5.1.0

# Embeddings (SIN torch para build más rápido)
sentence-transformers==3.2.1

# Database
PyMySQL==1.1.1
SQLAlchemy==2.0.35

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
cryptography==43.0.1

# Async & Queue
redis==5.2.0
celery==5.4.0

# Utilities
python-dotenv==1.0.1

# Testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0

# System monitoring dependencies
psutil>=5.9.0
GPUtil>=1.4.0

# Stress testing dependencies
aiohttp>=3.9.0

# Excel export
openpyxl>=3.1.0